JobVacancy
Most significant changes in Controller & Tests.

In `projects` table the `manager` column has been renamed to `manager_id` so that a distinction is made between $project->manager_id and $project->manager; the former merely returns a User id while the latter returns a User model instance.

We can now access the `manager` User instance from our job listing like so:
`$jobVacancy->job->project->manager`

Slug attribute has been temporarily removed. Wasn't really a slug tbh but a UUID hash. For the time being we will use the id as the resource locator for JobVacancy. We might potentially reintroduce a slug attribute to this model.

Create, Update & Delete operations are authorized by the `isManager` & `canEditDeleteJobListing` gates. Fixes security bug in JobVacancyController@store so that now users can only create a JobVacancy for projects which they are managing.

Factory & Seeders updated to generate more realistic dummy data as well as limiting the locations of these jobs to a small subset of popular cities so that we can later more easily test out API's filtering capabilities.

We have also set up some Eloquent relationships such as: `$city->jobVacancies`

`isOwnerOfSkill` gate has been deleted as no longer necessary.

+++ app/Http/Controllers/JobVacancyController.php
@@ -3,7 +3,9 @@
 namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
+use Illuminate\Support\Facades\Gate;
 
+use App\Models\Job;
 use App\Models\JobVacancy;
 
 class JobVacancyController extends Controller
@@ -33,8 +35,11 @@ class JobVacancyController extends Controller
 
         ]);
 
+        $project = Job::find($request['job_id'])->project;
+
+        Gate::authorize('isManager', $project);
+
         $data = $request->all();
-        $data['slug'] = md5(uniqid(rand(), true));
         $data['ends_at'] = date('Y-m-d', strtotime("+30 day", strtotime(now())));
 
         return JobVacancy::create($data);
@@ -64,9 +69,14 @@ class JobVacancyController extends Controller
     {
         // validate?
 
-        $vacancy = JobVacancy::find($id);
-        $vacancy->update($request->all());
-        return $vacancy;
+        $listing = JobVacancy::find($id);
+        $project = $listing->job->project;
+
+        Gate::authorize('canEditDeleteJobListing', $listing);
+
+        $listing->update($request->all());
+
+        return $listing;
     }
 
     /**
@@ -75,11 +85,16 @@ class JobVacancyController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
-    public function destroy($id)
+    public function destroy(Request $request, $id)
     {
-        // use middleware or just check user_id
+        $listing = JobVacancy::findOrFail($id);
+        $project = $listing->job->project;
 
-        return JobVacancy::destroy($id);
+        Gate::authorize('canEditDeleteJobListing', $listing);
+
+        JobVacancy::destroy($id);
+
+        return response('Job Listing has been deleted', 200);
     }
 
     /**
